feat: add script to search Slack messages programmatically
This commit adds a script that allows users to search for messages in Slack programmatically. The script uses the Slack API to perform the search and handles pagination to retrieve all matching messages. It also handles rate limiting by waiting for the rate limit to be lifted before resuming the search. 

The script prompts the user to enter the query they want to search for and provides an option to save the results to a file. If the user chooses to save the results, they are prompted to enter the name of the file. The script then displays the number of results found and, if a file name was provided, saves the results to the specified file.

This script provides a convenient way to search for messages in Slack without having to use the Slack interface.